+++ src/main/java/io/github/astrapi69/io/file/filter/TxtFileFilter.java
@@ -36,11 +36,8 @@ import io.github.astrapi69.io.file.FileExtension;
  * @version 1.0
  * @author Asterios Raptis
  */
-public class TxtFileFilter implements FileFilter
+public class TxtFileFilter extends SuffixFileFilter
 {
-
-	/** The recursive flag */
-	private final boolean recursive;
 
 	/**
 	 * Instantiates a new {@link TxtFileFilter} with the given flag for recursion
@@ -50,7 +47,7 @@ public class TxtFileFilter implements FileFilter
 	 */
 	public TxtFileFilter(final boolean recursive)
 	{
-		this.recursive = recursive;
+		super(FileExtension.TXT.getExtension(), recursive);
 	}
 
 	/**
@@ -59,28 +56,6 @@ public class TxtFileFilter implements FileFilter
 	public TxtFileFilter()
 	{
 		this(true);
-	}
-
-	/**
-	 * {@inheritDoc}
-	 */
-	@Override
-	public boolean accept(final File file)
-	{
-		// if recursive flag true and file is a directory allow recursive search
-		if (recursive && file.isDirectory())
-		{
-			return true;
-		}
-		else
-		{
-			final String fileName = file.getName().toLowerCase();
-			if (fileName.endsWith(FileExtension.TXT.getExtension()))
-			{
-				return true;
-			}
-		}
-		return false;
 	}
 
 	/**